+++ src/Symfony/Bundle/SecurityBundle/DependencyInjection/MainConfiguration.php
@@ -87,6 +87,7 @@ class MainConfiguration implements ConfigurationInterface
     {
         $rootNode
             ->children()
+                ->scalarNode('prng_seed_provider')->defaultNull()->end()
                 ->arrayNode('acl')
                     ->children()
                         ->scalarNode('connection')